Added team to user and item models/controllers/authentication. Added a new data model for Agenda items along with the necessary HTML/controllers/routers/authentication.

+++ controllers/items.js
@@ -1,22 +1,24 @@
 const { StatusCodes } = require("http-status-codes");
 const Item = require("../models/Item");
+const User = require("../models/User");
 const { BadRequestError, NotFoundError } = require("../errors");
 
 const getAllItems = async (req, res) => {
-  const items = await Item.find({ createdBy: req.user.userId }).sort(
-    "createdAt"
-  );
+  const user = await User.find({ _id: req.user.userId });
+  const team = user[0].team;
+  const items = await Item.find({ team: team }).sort("createdAt");
   res.status(StatusCodes.OK).json({ items, count: items.length });
 };
 
 const getSingleItem = async (req, res) => {
+  const user = await User.find({ _id: req.user.userId });
+  const team = user[0].team;
   const {
-    user: { userId },
     params: { id: itemId },
   } = req;
   const item = await Item.findOne({
     _id: itemId,
-    createdBy: userId,
+    team: team,
   });
   if (!item) {
     throw new NotFoundError(`No Item matching id ${itemId}`);
@@ -25,15 +27,18 @@ const getSingleItem = async (req, res) => {
 };
 
 const createItem = async (req, res) => {
+  const user = await User.find({ _id: req.user.userId });
   req.body.createdBy = req.user.userId;
+  req.body.team = user[0].team;
   const item = await Item.create(req.body);
   res.status(StatusCodes.CREATED).json({ item });
 };
 
 const updateItem = async (req, res) => {
+  const user = await User.find({ _id: req.user.userId });
+  const team = user[0].team;
   const {
     body: { title, value },
-    user: { userId },
     params: { id: itemId },
   } = req;
   if (!title || !value) {
@@ -42,7 +47,7 @@ const updateItem = async (req, res) => {
   const item = await Item.findByIdAndUpdate(
     {
       _id: itemId,
-      createdBy: userId,
+      team: team,
     },
     req.body,
     { new: true, runValidators: true }
@@ -54,18 +59,31 @@ const updateItem = async (req, res) => {
 };
 
 const deleteItem = async (req, res) => {
+  const user = await User.find({ _id: req.user.userId });
+  const userTeam = user[0].team;
   const {
-    user: { userId },
     params: { id: itemId },
   } = req;
-  const item = await Item.findByIdAndDelete({
+  const item = await Item.find({
     _id: itemId,
-    createdBy: userId,
   });
-  if (!item) {
-    throw new NotFoundError(`No Item matching id ${itemId}`);
+  const itemTeam = item[0].team;
+
+  if (userTeam !== itemTeam) {
+    throw new BadRequestError(
+      "You may not delete an item from another teams budget."
+    );
   }
-  res.status(StatusCodes.OK).json({ msg: `Item ${itemId} deleted` });
+  if (!item) {
+    throw new NotFoundError(
+      `No Item matching id ${itemId}. It may have already been deleted. Try refreshing.`
+    );
+  } else {
+    const deletedItem = await Item.findByIdAndDelete({
+      _id: itemId,
+    });
+    res.status(StatusCodes.OK).json({ msg: `Item ${deletedItem} deleted` });
+  }
 };
 
 module.exports = {
